make edits and split into two files
